requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
mysqlclient==2.1.1
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
